+++ LaunchServer/src/main/java/ru/gravit/launchserver/LaunchServer.java
@@ -330,7 +330,10 @@ public final class LaunchServer implements Runnable {
     public LaunchServer(Path dir, String[] args) throws IOException, InvalidKeySpecException {
         this.dir = dir;
         launcherLibraries = dir.resolve("launcher-libraries");
-        if (!Files.isDirectory(launcherLibraries)) Files.createDirectory(launcherLibraries);
+        if (!Files.isDirectory(launcherLibraries)) {
+        	Files.deleteIfExists(launcherLibraries);
+        	Files.createDirectory(launcherLibraries);
+        }
         this.args = Arrays.asList(args);
         configFile = dir.resolve("LaunchServer.conf");
         publicKeyFile = dir.resolve("public.key");
