+++ Launcher/build.gradle
@@ -1,19 +1,17 @@
 String mainClassName = "ru.gravit.launcher.ClientLauncherWrapper"
 String mainAgentName = "ru.gravit.launcher.LauncherAgent"
 
-
 repositories {
-    maven {
-        url "http://repo.spring.io/plugins-release/"
-    }
+	maven {
+		url "http://repo.spring.io/plugins-release/"
+	}
 }
 
 sourceCompatibility = '1.8'
 targetCompatibility = '1.8'
 
 jar {
-    from { configurations.runtime.collect { it.isDirectory() ? it : zipTree(it) } }
-    manifest.attributes("Main-Class": mainClassName,
+	manifest.attributes("Main-Class": mainClassName,
 		"Premain-Class": mainAgentName,
 		"Can-Redefine-Classes": "true",
 		"Can-Retransform-Classes": "true",
@@ -21,14 +19,20 @@ jar {
 }
 
 dependencies {
-    compile project(':LauncherAPI')
-    compile 'com.github.oshi:oshi-core:3.11.0'
+	compile project(':LauncherAPI')
+	compile 'com.github.oshi:oshi-core:3.11.0'
 }
 
 task genRuntimeJS(type: Zip) {
-    archiveName = "runtime.zip"
-    destinationDir = file("${buildDir}/tmp")
-    from "runtime/"
+	archiveName = "runtime.zip"
+	destinationDir = file("${buildDir}/tmp")
+	from "runtime/"
 }
 
-build.dependsOn tasks.genRuntimeJS
+task dumpLibs(type: Copy) {
+	into "$buildDir/libs/libraries"
+	from configurations.runtime
+}
+
+
+build.dependsOn tasks.genRuntimeJS, tasks.dumpLibs
