+++ LauncherAPI/build.gradle
@@ -2,8 +2,8 @@ sourceCompatibility = '1.8'
 targetCompatibility = '1.8'
 
 dependencies {
-    compile project(':libLauncher')
-    compile 'javax.websocket:javax.websocket-client-api:1.1'
-    compileOnly 'com.google.guava:guava:26.0-jre'
-    compile files('../compat/authlib/authlib-clean.jar')
+	compile project(':libLauncher')
+	compile 'javax.websocket:javax.websocket-client-api:1.1'
+	compileOnly 'com.google.guava:guava:26.0-jre'
+	compile files('../compat/authlib/authlib-clean.jar')
 }
