+++ ServerWrapper/src/main/java/ru/gravit/launcher/server/ServerWrapper.java
@@ -1,6 +1,5 @@
 package ru.gravit.launcher.server;
 
-
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import ru.gravit.launcher.ClientPermissions;
